refactor(stylexswc/rs-compiler): hold the fixture vocabulary apart from the addon
`lib/types.ts` kept the names a fixture is described with beside a value
it reads off `dist/index.js`. Importing that module loads this package's
native binding, so a process that must hold no binding but its subject's
could not ask for the names, and had to carry a copy of them.

`lib/fixture-schema.ts` holds them now, and nothing in it needs the addon.
It carries the weights, the categories, the option keys, the `sourceMap`
spellings, and the reader that narrows a fixture's overrides key by key.
`lib/types.ts` passes every name on, so each module that already asked it
for one keeps asking it.

The override reader takes the one step its two callers do not share. The
manifest reader looks a spelling up in the package's own `SourceMaps`
export. A reader without the addon cannot, and supplies its own step.

Two helpers move to `lib/json.ts`, which says that every parser in `lib/`
funnels through it so one rule cannot drift between schemas.
`requireOneOf` was already written twice, and the two copies had begun to
say different things about the same fault.
